+++ app/src/main/java/com/dhy/coffeesecret/MainActivity.java
@@ -10,15 +10,15 @@ import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.ImageView;
 
-import com.dhy.coffeesecret.ui.mycontainer.MyContainerFragment;
-import com.dhy.coffeesecret.ui.mycup.MyCupFragment;
-import com.dhy.coffeesecret.ui.mydevice.MyDeviceFragment;
+import com.dhy.coffeesecret.ui.container.ContainerFragment;
+import com.dhy.coffeesecret.ui.cup.CupFragment;
+import com.dhy.coffeesecret.ui.device.DeviceFragment;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class MainActivity extends AppCompatActivity implements MyDeviceFragment.OnDeviceInteractionListener,
-        MyContainerFragment.OnContainerInteractionListener, MyCupFragment.OnCupInteractionListener{
+public class MainActivity extends AppCompatActivity implements DeviceFragment.OnDeviceInteractionListener,
+        ContainerFragment.OnContainerInteractionListener, CupFragment.OnCupInteractionListener{
 
     // 标签页
     private TabLayout mTabLayout;
@@ -26,8 +26,8 @@ public class MainActivity extends AppCompatActivity implements MyDeviceFragment.
     private int[] imgSelectorIds = {R.drawable.nav_device_selector, R.drawable.nav_container_selector, R.drawable.nav_cup_selector};
     // 滑动页面视图
     private ViewPager mViewPager;
-    // fragment集合
-    private List<Fragment> mFragments = new ArrayList<>();
+    // fragment集合 //
+    private Fragment[] mFragments;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -42,9 +42,11 @@ public class MainActivity extends AppCompatActivity implements MyDeviceFragment.
      */
     private void initParam(){
         // 初始化fragment视图
-        mFragments.add(new MyDeviceFragment());
-        mFragments.add(new MyContainerFragment());
-        mFragments.add(new MyCupFragment());
+
+        mFragments =new Fragment[] {new DeviceFragment(),new ContainerFragment(),new CupFragment()};
+//        mFragments.add(new DeviceFragment());
+//        mFragments.add(new ContainerFragment());
+//        mFragments.add(new CupFragment());
 
         // 获取id
         mTabLayout = (TabLayout)findViewById(R.id.id_fragment_tabLayout);
@@ -54,45 +56,55 @@ public class MainActivity extends AppCompatActivity implements MyDeviceFragment.
         mViewPager.setAdapter(new FragmentPagerAdapter(getSupportFragmentManager()) {
             @Override
             public Fragment getItem(int position) {
-                return mFragments.get(position);
+                return mFragments[position];
             }
 
             @Override
             public int getCount() {
-                return mFragments.size();
+                return mFragments.length;
             }
         });
-        // 默认页面滑动监听器
-        mViewPager.addOnPageChangeListener(new TabLayout.TabLayoutOnPageChangeListener(mTabLayout));
-        // 设置首页
-        mViewPager.setCurrentItem(0);
 
+        mTabLayout.setupWithViewPager(mViewPager);
+
+        mTabLayout.getTabAt(0).setCustomView(getCustomerView(0));
+        mTabLayout.getTabAt(1).setCustomView(getCustomerView(1));
+        mTabLayout.getTabAt(2).setCustomView(getCustomerView(2));
+        mViewPager.setCurrentItem(0);
         // 设置tablayout固定
         mTabLayout.setTabMode(TabLayout.MODE_FIXED);
-        // 设置tablayout布局布满
-        mTabLayout.setTabGravity(TabLayout.GRAVITY_FILL);
-        // 添加新的三个标签页
-        mTabLayout.addTab(mTabLayout.newTab().setCustomView(getCustomerView(0)));
-        mTabLayout.addTab(mTabLayout.newTab().setCustomView(getCustomerView(1)));
-        mTabLayout.addTab(mTabLayout.newTab().setCustomView(getCustomerView(2)));
 
-        // 监听标签页单击事件
-        mTabLayout.addOnTabSelectedListener(new TabLayout.OnTabSelectedListener() {
-            @Override
-            public void onTabSelected(TabLayout.Tab tab) {
-                mViewPager.setCurrentItem(tab.getPosition(), true);
-            }
-
-            @Override
-            public void onTabUnselected(TabLayout.Tab tab) {
-
-            }
-
-            @Override
-            public void onTabReselected(TabLayout.Tab tab) {
-
-            }
-        });
+//        // 默认页面滑动监听器
+//        mViewPager.addOnPageChangeListener(new TabLayout.TabLayoutOnPageChangeListener(mTabLayout));
+//        // 设置首页
+//        mViewPager.setCurrentItem(0);
+//
+//        // 设置tablayout固定
+//        mTabLayout.setTabMode(TabLayout.MODE_FIXED);
+//        // 设置tablayout布局布满
+//        mTabLayout.setTabGravity(TabLayout.GRAVITY_FILL);
+//        // 添加新的三个标签页
+//        mTabLayout.addTab(mTabLayout.newTab().setCustomView(getCustomerView(0)));
+//        mTabLayout.addTab(mTabLayout.newTab().setCustomView(getCustomerView(1)));
+//        mTabLayout.addTab(mTabLayout.newTab().setCustomView(getCustomerView(2)));
+//
+//        // 监听标签页单击事件
+//        mTabLayout.addOnTabSelectedListener(new TabLayout.OnTabSelectedListener() {
+//            @Override
+//            public void onTabSelected(TabLayout.Tab tab) {
+//                mViewPager.setCurrentItem(tab.getPosition(), true);
+//            }
+//
+//            @Override
+//            public void onTabUnselected(TabLayout.Tab tab) {
+//
+//            }
+//
+//            @Override
+//            public void onTabReselected(TabLayout.Tab tab) {
+//
+//            }
+//        });
 
     }
