+++ app/src/main/java/com/dhy/coffeesecret/ui/cup/CupFragment.java
@@ -1,4 +1,4 @@
-package com.dhy.coffeesecret.ui.mydevice;
+package com.dhy.coffeesecret.ui.cup;
 
 import android.content.Context;
 import android.net.Uri;
@@ -14,12 +14,12 @@ import com.dhy.coffeesecret.R;
 /**
  * A simple {@link Fragment} subclass.
  * Activities that contain this fragment must implement the
- * {@link OnDeviceInteractionListener} interface
+ * {@link OnCupInteractionListener} interface
  * to handle interaction events.
- * Use the {@link MyDeviceFragment#newInstance} factory method to
+ * Use the {@link CupFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class MyDeviceFragment extends Fragment {
+public class CupFragment extends Fragment {
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -29,9 +29,9 @@ public class MyDeviceFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    private OnDeviceInteractionListener mListener;
+    private OnCupInteractionListener mListener;
 
-    public MyDeviceFragment() {
+    public CupFragment() {
         // Required empty public constructor
     }
 
@@ -41,11 +41,11 @@ public class MyDeviceFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment MyDeviceFragment.
+     * @return A new instance of fragment CupFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static MyDeviceFragment newInstance(String param1, String param2) {
-        MyDeviceFragment fragment = new MyDeviceFragment();
+    public static CupFragment newInstance(String param1, String param2) {
+        CupFragment fragment = new CupFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -60,7 +60,7 @@ public class MyDeviceFragment extends Fragment {
             mParam1 = getArguments().getString(ARG_PARAM1);
             mParam2 = getArguments().getString(ARG_PARAM2);
         }
-        Log.e("codelevex", "33333");
+        Log.e("codelevex", "22222");
 
     }
 
@@ -68,21 +68,21 @@ public class MyDeviceFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_my_device, container, false);
+        return inflater.inflate(R.layout.fragment_my_cup, container, false);
     }
 
     // TODO: Rename method, update argument and hook method into UI event
     public void onButtonPressed(Uri uri) {
         if (mListener != null) {
-            mListener.onDeviceInteraction(uri);
+            mListener.onCupInteraction(uri);
         }
     }
 
     @Override
     public void onAttach(Context context) {
         super.onAttach(context);
-        if (context instanceof OnDeviceInteractionListener) {
-            mListener = (OnDeviceInteractionListener) context;
+        if (context instanceof OnCupInteractionListener) {
+            mListener = (OnCupInteractionListener) context;
         } else {
             throw new RuntimeException(context.toString()
                     + " must implement OnContainerInteractionListener");
@@ -105,8 +105,8 @@ public class MyDeviceFragment extends Fragment {
      * "http://developer.android.com/training/basics/fragments/communicating.html"
      * >Communicating with Other Fragments</a> for more information.
      */
-    public interface OnDeviceInteractionListener {
+    public interface OnCupInteractionListener {
         // TODO: Update argument type and name
-        void onDeviceInteraction(Uri uri);
+        void onCupInteraction(Uri uri);
     }
 }
