+++ app/src/main/res/layout/activity_main.xml
@@ -13,7 +13,6 @@
         android:id="@+id/id_fragment_tabLayout"
         android:layout_width="match_parent"
         android:layout_height="55dp"
-
         app:tabIndicatorColor="@color/tab_underline"
         app:tabSelectedTextColor="@color/tab_selected"
         app:tabTextColor="@color/tab_default">
@@ -28,6 +27,5 @@
         android:id="@+id/id_fragment_viewPager"
         android:layout_width="match_parent"
         android:layout_height="match_parent">
-
     </android.support.v4.view.ViewPager>
 </RelativeLayout>
