+++ app/src/main/res/layout/fragment_my_container.xml
@@ -2,7 +2,7 @@
              xmlns:tools="http://schemas.android.com/tools"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
-             tools:context=".ui.mycontainer.MyContainerFragment">
+             tools:context=".ui.container.ContainerFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
