+++ app/src/main/res/layout/fragment_my_device.xml
@@ -2,7 +2,7 @@
              xmlns:tools="http://schemas.android.com/tools"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
-             tools:context=".ui.mydevice.MyDeviceFragment">
+             tools:context=".ui.device.DeviceFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
